+++ utils/getCurrentGoal.ts
@@ -1,7 +1,7 @@
 import type { Goal } from '@/doar-para';
 
 export default (goals:Goal[], totalAmount:number = 0) => {
-  const sortedGoals = [...goals].sort((a, b) => a.amount - b.amount);
+  const sortedGoals = goals.slice().sort((a, b) => a.amount - b.amount);
 
   return (
     sortedGoals.find((x: Goal) => x.amount > totalAmount)