+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - perf: Simplify calculating current goal
 - perf: Load testimonies on client side only
+- perf: Simplify donations summary
 
 ## v3.4.12 - 2026-03-20